First sprint refactoring
