+++ src/main/java/at/fhtw/swen3/Application.java
@@ -1,4 +1,4 @@
-package org.openapitools;
+package at.fhtw.swen3;
 
 import com.fasterxml.jackson.databind.Module;
 import org.openapitools.jackson.nullable.JsonNullableModule;
@@ -8,11 +8,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"org.openapitools", "at.fhtw.swen3.services" , "org.openapitools.configuration"})
-public class OpenApiGeneratorApplication {
+@ComponentScan(basePackages = {"at.fhtw.swen3.services", "at.fhtw.swen3.configuration"})
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(OpenApiGeneratorApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
     @Bean
