+++ src/main/java/at/fhtw/swen3/configuration/SpringDocConfiguration.java
@@ -1,4 +1,4 @@
-package org.openapitools.configuration;
+package at.fhtw.swen3.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,13 +6,9 @@ import org.springframework.context.annotation.Configuration;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.Contact;
-import io.swagger.v3.oas.models.info.License;
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.security.SecurityScheme;
 
 @Configuration
 public class SpringDocConfiguration {
-
     @Bean
     OpenAPI apiInfo() {
         return new OpenAPI()
