+++ src/main/java/at/fhtw/swen3/persistence/Warehouse.java
@@ -1,26 +1,16 @@
 package at.fhtw.swen3.persistence;
 
-import java.net.URI;
 import java.util.Objects;
-import at.fhtw.swen3.persistence.GeoCoordinate;
-import at.fhtw.swen3.persistence.Hop;
-import at.fhtw.swen3.persistence.WarehouseAllOfNextHops;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import java.util.ArrayList;
 import java.util.List;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
 import javax.validation.Valid;
 import javax.validation.constraints.*;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 
-import java.util.*;
 import javax.annotation.Generated;
 
 /**
@@ -37,7 +27,7 @@ public class Warehouse extends Hop {
 
   @JsonProperty("nextHops")
   @Valid
-  private List<WarehouseAllOfNextHops> nextHops = new ArrayList<>();
+  private List<WarehouseNextHops> nextHops = new ArrayList<>();
 
   public Warehouse level(Integer level) {
     this.level = level;
@@ -58,12 +48,12 @@ public class Warehouse extends Hop {
     this.level = level;
   }
 
-  public Warehouse nextHops(List<WarehouseAllOfNextHops> nextHops) {
+  public Warehouse nextHops(List<WarehouseNextHops> nextHops) {
     this.nextHops = nextHops;
     return this;
   }
 
-  public Warehouse addNextHopsItem(WarehouseAllOfNextHops nextHopsItem) {
+  public Warehouse addNextHopsItem(WarehouseNextHops nextHopsItem) {
     this.nextHops.add(nextHopsItem);
     return this;
   }
@@ -74,11 +64,11 @@ public class Warehouse extends Hop {
   */
   @NotNull @Valid 
   @Schema(name = "nextHops", description = "Next hops after this warehouse (warehouses or trucks).", required = true)
-  public List<WarehouseAllOfNextHops> getNextHops() {
+  public List<WarehouseNextHops> getNextHops() {
     return nextHops;
   }
 
-  public void setNextHops(List<WarehouseAllOfNextHops> nextHops) {
+  public void setNextHops(List<WarehouseNextHops> nextHops) {
     this.nextHops = nextHops;
   }
