+++ src/main/java/at/fhtw/swen3/services/impl/HomeController.java
@@ -1,10 +1,7 @@
-package org.openapitools.configuration;
+package at.fhtw.swen3.services.impl;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.GetMapping;
 
 /**
  * Home redirection to OpenAPI api documentation
