+++ src/main/java/at/fhtw/swen3/services/impl/ParcelApiController.java
@@ -1,0 +1,28 @@
+package at.fhtw.swen3.services.impl;
+
+
+import at.fhtw.swen3.services.ParcelApi;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.context.request.NativeWebRequest;
+
+import java.util.Optional;
+import javax.annotation.Generated;
+
+@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-10-13T12:19:08.753753Z[Etc/UTC]")
+@Controller
+public class ParcelApiController implements ParcelApi {
+
+    private final NativeWebRequest request;
+
+    @Autowired
+    public ParcelApiController(NativeWebRequest request) {
+        this.request = request;
+    }
+
+    @Override
+    public Optional<NativeWebRequest> getRequest() {
+        return Optional.ofNullable(request);
+    }
+
+}
