+++ src/main/java/at/fhtw/swen3/util/ApiUtil.java
@@ -1,4 +1,4 @@
-package at.fhtw.swen3.services;
+package at.fhtw.swen3.util;
 
 import org.springframework.web.context.request.NativeWebRequest;
