+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8080
-spring.jackson.date-format=org.openapitools.RFC3339DateFormat
+spring.jackson.date-format=at.fhtw.swen3.util.RFC3339DateFormat
 spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
